==> 1_data_load.ipynb <==
import pandas as pd
import numpy as np
import yfinance as yf

# df is the main dataset used in the project
df_news = pd.read_csv("https://raw.githubusercontent.com/EagleAdelaide/FinSen_Dataset/refs/heads/main/data.pptx/FinSen_US_Categorized_Timestamp.csv")
df_news.head(10)

# NASDAQ data
nasdaq = yf.Ticker("^IXIC")
df_nasdaq = nasdaq.history(start="2007-04-01", end="2023-07-30")
df_nasdaq.head(10)

sp = yf.Ticker("^SPX")
df_sp = sp.history(start="2007-04-01", end="2023-07-30")
df_sp.head(10)

df_main = yf.download(['^IXIC', '^SPX'], period='1mo').head(10)
list(df_main.columns.values)
