==> 2.2_bayes.ipynb <==
%run 2.1_data_preparation.ipynb

from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import classification_report

tokenized_df['text'] = tokenized_df['Cleaned Content'].apply(lambda x: ' '.join(x)) 

vectorizer = CountVectorizer()
X = vectorizer.fit_transform(tokenized_df['text'])

y = tokenized_df['Label'] 

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

model = MultinomialNB()
model.fit(X_train, y_train)

y_pred = model.predict(X_test)

print(classification_report(y_test, y_pred))

#tokenized_df.head(10)
