==> 4.3_ranfor.ipynb <==
%run 2_data_preparation_news.ipynb

from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report
from sklearn.preprocessing import StandardScaler
from sklearn.feature_extraction.text import TfidfVectorizer
from scipy.sparse import hstack
from sklearn.model_selection import train_test_split
import pandas as pd

numeric_features = ['Prev_change', 'Sma_5', 'Volatility' ]

for col in ['Prev_change', 'Sma_5', 'Volatility' ]:
    tokenized_df[col] = tokenized_df[col].fillna(tokenized_df[col].mean())
    
numeric_data = tokenized_df[numeric_features]
scaler = StandardScaler()
X_numeric = scaler.fit_transform(numeric_data)

vectorizer = TfidfVectorizer(ngram_range=(1, 2), max_df=0.9, min_df=5)
X_text = vectorizer.fit_transform(tokenized_df['Content'])

X_combined = hstack([X_text, X_numeric])
y = tokenized_df['Simple Label']

X_train, X_test, y_train, y_test = train_test_split(X_combined, y, test_size=0.2, random_state=42)

rf_model = RandomForestClassifier(n_estimators=100, class_weight='balanced', random_state=42)
rf_model.fit(X_train, y_train)

y_pred_rf = rf_model.predict(X_test)
print(classification_report(y_test, y_pred_rf))

import matplotlib.pyplot as plt

feature_importances = rf_model.feature_importances_

feature_names = vectorizer.get_feature_names_out().tolist() + numeric_features
feature_importance_df = pd.DataFrame({'Feature': feature_names, 'Importance': feature_importances})

feature_importance_df = feature_importance_df.sort_values(by='Importance', ascending=False)

print(feature_importance_df)

feature_importance_df = feature_importance_df.sort_values(by='Importance', ascending=False).head(10)

plt.figure(figsize=(10, 6))
plt.barh(feature_importance_df['Feature'], feature_importance_df['Importance'])
plt.xlabel('Importance')
plt.title('Feature Importance (Random Forest)')
plt.gca().invert_yaxis()
plt.show()
